refactor(PlayTimeValidator, CarNameValidator): 싱글톤으로 바꾸기
